// Front/calitel-front/src/app/enums/telephone-endpoints.enum.ts
export enum TelephoneEndpoints {
  TelephoneDomain = "/telephoneNumber",
  TelephoneByCustomer = "/telephoneNumber/customer",
  TelephoneTrackingProcess = "/telephoneNumber/trackingProcess",
  AssignTelephone = "/telephoneNumber/assign",
  ReleaseTelephone = "/telephoneNumber/release",
  DownloadCustomerHistory = "/telephoneNumber/downloadCustomerHistory",
  DownloadTelephoneHistory = "/telephoneNumber/downloadNumberHistory",
  TimeSetting = "/telephoneNumber/timeSetting",
}